Move input state parsing code into separate file.
